🚚 rename: my-page 파일명 수정
